+++ src/jenn/utilities/__init__.py
@@ -3,7 +3,8 @@
 
 from ._finite_difference import finite_difference
 from ._jmp import from_jmp
+from ._load import load_csv, load_npz
 from ._rbf import rbf
 from ._sample import sample
 
-__all__ = ["finite_difference", "from_jmp", "rbf", "sample"]
+__all__ = ["finite_difference", "from_jmp", "load_csv", "load_npz", "rbf", "sample"]